Remove the old agent writer
