+++ src/Datadog.Trace/Agent/IApi.cs
@@ -5,8 +5,6 @@ namespace Datadog.Trace.Agent
 {
     internal interface IApi
     {
-        Task<bool> SendTracesAsync(Span[][] traces);
-
         Task<bool> SendTracesAsync(ArraySegment<byte> traces, int numberOfTraces);
     }
 }
