+++ src/Datadog.Trace/Agent/IApiRequest.cs
@@ -1,14 +1,11 @@
 using System;
 using System.Threading.Tasks;
-using Datadog.Trace.Agent.MessagePack;
 
 namespace Datadog.Trace.Agent
 {
     internal interface IApiRequest
     {
         void AddHeader(string name, string value);
-
-        Task<IApiResponse> PostAsync(Span[][] traces, FormatterResolverWrapper formatterResolver);
 
         Task<IApiResponse> PostAsync(ArraySegment<byte> traces);
     }
