+++ src/Datadog.Trace/Agent/Transports/ApiWebRequest.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Net;
 using System.Threading.Tasks;
-using Datadog.Trace.Agent.MessagePack;
 
 namespace Datadog.Trace.Agent.Transports
 {
@@ -24,29 +23,6 @@ namespace Datadog.Trace.Agent.Transports
         public void AddHeader(string name, string value)
         {
             _request.Headers.Add(name, value);
-        }
-
-        public async Task<IApiResponse> PostAsync(Span[][] traces, FormatterResolverWrapper formatterResolver)
-        {
-            _request.Method = "POST";
-            _request.ContentType = "application/msgpack";
-
-            using (var requestStream = await _request.GetRequestStreamAsync().ConfigureAwait(false))
-            {
-                await CachedSerializer.Instance.SerializeAsync(requestStream, traces, formatterResolver).ConfigureAwait(false);
-            }
-
-            try
-            {
-                var httpWebResponse = (HttpWebResponse)await _request.GetResponseAsync().ConfigureAwait(false);
-                return new ApiWebResponse(httpWebResponse);
-            }
-            catch (WebException exception)
-                when (exception.Status == WebExceptionStatus.ProtocolError && exception.Response != null)
-            {
-                // If the exception is caused by an error status code, ignore it and let the caller handle the result
-                return new ApiWebResponse((HttpWebResponse)exception.Response);
-            }
         }
 
         public async Task<IApiResponse> PostAsync(ArraySegment<byte> traces)
