+++ src/Datadog.Trace/Agent/Transports/HttpClientRequest.cs
@@ -22,19 +22,6 @@ namespace Datadog.Trace.Agent.Transports
             _request.Headers.Add(name, value);
         }
 
-        public async Task<IApiResponse> PostAsync(Span[][] traces, FormatterResolverWrapper formatterResolver)
-        {
-            // re-create HttpContent on every retry because some versions of HttpClient always dispose of it, so we can't reuse.
-            using (var content = new TracesMessagePackContent(traces, formatterResolver))
-            {
-                _request.Content = content;
-
-                var response = await _client.SendAsync(_request).ConfigureAwait(false);
-
-                return new HttpClientResponse(response);
-            }
-        }
-
         public async Task<IApiResponse> PostAsync(ArraySegment<byte> traces)
         {
             // re-create HttpContent on every retry because some versions of HttpClient always dispose of it, so we can't reuse.
