+++ test/Datadog.Trace.Tests/AgentWriterTests.cs
@@ -1,7 +1,9 @@
 using System;
+using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
 using Datadog.Trace.Agent;
+using Datadog.Trace.Agent.MessagePack;
 using Moq;
 using Xunit;
 
@@ -29,16 +31,23 @@ namespace Datadog.Trace.Tests
         [Fact]
         public async Task WriteTrace_2Traces_SendToApi()
         {
-            // TODO:bertrand it is too complicated to setup such a simple test
-            var trace = new[] { new Span(_spanContext, start: null) };
-            _agentWriter.WriteTrace(trace);
-            await _agentWriter.FlushTracesAsync(); // Force a flush to make sure the trace is written to the API
-            _api.Verify(x => x.SendTracesAsync(It.Is<Span[][]>(y => y.Single().Equals(trace))), Times.Once);
+            var trace = new[] { new Span(new SpanContext(1, 1), DateTimeOffset.UtcNow) };
+            var expectedData1 = Vendors.MessagePack.MessagePackSerializer.Serialize(trace, new FormatterResolverWrapper(SpanFormatterResolver.Instance));
 
-            trace = new[] { new Span(_spanContext, start: null) };
             _agentWriter.WriteTrace(trace);
             await _agentWriter.FlushTracesAsync(); // Force a flush to make sure the trace is written to the API
-            _api.Verify(x => x.SendTracesAsync(It.Is<Span[][]>(y => y.Single().Equals(trace))), Times.Once);
+
+            _api.Verify(x => x.SendTracesAsync(It.Is<ArraySegment<byte>>(y => Equals(y, expectedData1)), It.Is<int>(i => i == 1)), Times.Once);
+
+            _api.ResetCalls();
+
+            trace = new[] { new Span(new SpanContext(2, 2), DateTimeOffset.UtcNow) };
+            var expectedData2 = Vendors.MessagePack.MessagePackSerializer.Serialize(trace, new FormatterResolverWrapper(SpanFormatterResolver.Instance));
+
+            _agentWriter.WriteTrace(trace);
+            await _agentWriter.FlushTracesAsync(); // Force a flush to make sure the trace is written to the API
+
+            _api.Verify(x => x.SendTracesAsync(It.Is<ArraySegment<byte>>(y => Equals(y, expectedData2)), It.Is<int>(i => i == 1)), Times.Once);
         }
 
         [Fact]
@@ -47,6 +56,11 @@ namespace Datadog.Trace.Tests
             var w = new AgentWriter(_api.Object, statsd: null);
             await w.FlushAndCloseAsync();
             await w.FlushAndCloseAsync();
+        }
+
+        private bool Equals(ArraySegment<byte> data, byte[] expectedData)
+        {
+            return data.Array.Skip(data.Offset).Take(data.Count).Skip(SpanBuffer.HeaderSize).SequenceEqual(expectedData);
         }
     }
 }
