+++ test/benchmarks/Benchmarks.Trace/AgentWriterBenchmark.cs
@@ -4,7 +4,6 @@ using System.Threading.Tasks;
 using BenchmarkDotNet.Attributes;
 using Datadog.Trace;
 using Datadog.Trace.Agent;
-using Datadog.Trace.Agent.MessagePack;
 using Datadog.Trace.Agent.Transports;
 using Datadog.Trace.BenchmarkDotNet;
 using Datadog.Trace.Configuration;
@@ -18,7 +17,6 @@ namespace Benchmarks.Trace
         private const int SpanCount = 1000;
 
         private static readonly IAgentWriter AgentWriter;
-        private static readonly IAgentWriter EagerAgentWriter;
         private static readonly Span[] Spans;
         private static readonly Span[] EnrichedSpans;
 
@@ -31,8 +29,7 @@ namespace Benchmarks.Trace
 
             var api = new Api(settings.AgentUri, new FakeApiRequestFactory(), statsd: null);
 
-            AgentWriter = new AgentWriter(api, statsd: null, automaticFlush: false, queueSize: SpanCount * 2);
-            EagerAgentWriter = new EagerAgentWriter(api, statsd: null, automaticFlush: false);
+            AgentWriter = new AgentWriter(api, statsd: null, automaticFlush: false);
 
             Spans = new Span[SpanCount];
             EnrichedSpans = new Span[SpanCount];
@@ -48,9 +45,7 @@ namespace Benchmarks.Trace
 
             // Run benchmarks once to reduce noise
             new AgentWriterBenchmark().WriteAndFlushTraces().GetAwaiter().GetResult();
-            new AgentWriterBenchmark().WriteAndFlushTracesNew().GetAwaiter().GetResult();
             new AgentWriterBenchmark().WriteAndFlushEnrichedTraces().GetAwaiter().GetResult();
-            new AgentWriterBenchmark().WriteAndFlushEnrichedTracesNew().GetAwaiter().GetResult();
         }
 
         /// <summary>
@@ -63,13 +58,6 @@ namespace Benchmarks.Trace
             return AgentWriter.FlushTracesAsync();
         }
 
-        [Benchmark]
-        public Task WriteAndFlushTracesNew()
-        {
-            EagerAgentWriter.WriteTrace(Spans);
-            return EagerAgentWriter.FlushTracesAsync();
-        }
-
         /// <summary>
         /// Same as WriteAndFlushTraces but with more realistic traces (with tags and metrics)
         /// </summary>
@@ -78,13 +66,6 @@ namespace Benchmarks.Trace
         {
             AgentWriter.WriteTrace(EnrichedSpans);
             return AgentWriter.FlushTracesAsync();
-        }
-
-        [Benchmark]
-        public Task WriteAndFlushEnrichedTracesNew()
-        {
-            EagerAgentWriter.WriteTrace(EnrichedSpans);
-            return EagerAgentWriter.FlushTracesAsync();
         }
 
         /// <summary>
@@ -120,16 +101,6 @@ namespace Benchmarks.Trace
             public void AddHeader(string name, string value)
             {
                 _realRequest.AddHeader(name, value);
-            }
-
-            public async Task<IApiResponse> PostAsync(Span[][] traces, FormatterResolverWrapper formatterResolver)
-            {
-                using (var requestStream = Stream.Null)
-                {
-                    await CachedSerializer.Instance.SerializeAsync(requestStream, traces, formatterResolver).ConfigureAwait(false);
-                }
-
-                return new FakeApiResponse();
             }
 
             public async Task<IApiResponse> PostAsync(ArraySegment<byte> traces)
